Remove focus from search field on app launch
Bringing up the app drawer gives focus to the search field. This focus is never removed when we launch an app,
causing the keyboard to come up when we return to the main launcher view as described in #38.

Added a call to clearFocus() on the search field in onAppClicked(), removing the focus from the search field
and allowing us to return to the launcher without opening the keyboard as expected.

Closes #38